requirements.txt
grpcio==1.67.1
grpcio-tools==1.67.1
protobuf==5.28.3
Flask==3.0.3
